Add a diff service to Overlaybd snapshotter (#21)
This diff service identifies whether the Compare function is operating
on overlaybd devices, and if so, attempts to call overlaybd-commit on
the writable_data and writable_index of the upper layer, efficiently
generating a blob for the writable layer.

This is much faster than the default diff service, which mounts the
block devices and compares the filesystem's file-by-file, which
generates an OCI tar which we then convert to overlaybd.